refactor: move authToken validation and useKeyPress to middlewares
