+++ src/stores/index.ts
@@ -4,5 +4,4 @@ export const createPinia = () => {
   return _createPinia();
 };
 
-export * from './boards/store';
 export * from './store';
